Casts in addition corrected
--HG--
branch : issue/HAST-136

+++ Posit/Posit16_0.cs
@@ -824,14 +824,14 @@ namespace Lombiq.Arithmetics
 			if (resultFractionBits == 0) return new Posit16_0(0, true);
 
 			var resultRegimeKValue = scaleFactor / (1 << MaximumExponentSize);
-			var resultExponentBits = (ushort)(scaleFactor % (1 << MaximumExponentSize));
+			var resultExponentBits = (scaleFactor % (1 << MaximumExponentSize));
 			if (resultExponentBits < 0)
 			{
 				resultRegimeKValue -= 1;
-				resultExponentBits += (ushort)(1 << MaximumExponentSize);
+				resultExponentBits += (1 << MaximumExponentSize);
 			}
 
-			return new Posit16_0(AssemblePositBitsWithRounding(resultSignBit, resultRegimeKValue, 0, resultFractionBits), true);
+			return new Posit16_0(AssemblePositBitsWithRounding(resultSignBit, resultRegimeKValue,(ushort) 0, resultFractionBits), true);
 		}
 
 		public static Posit16_0 operator +(Posit16_0 left, int right) => left + new Posit16_0(right);
