+++ Posit/Posit8_0.cs
@@ -824,14 +824,14 @@ namespace Lombiq.Arithmetics
 			if (resultFractionBits == 0) return new Posit8_0(0, true);
 
 			var resultRegimeKValue = scaleFactor / (1 << MaximumExponentSize);
-			var resultExponentBits = (byte)(scaleFactor % (1 << MaximumExponentSize));
+			var resultExponentBits = (scaleFactor % (1 << MaximumExponentSize));
 			if (resultExponentBits < 0)
 			{
 				resultRegimeKValue -= 1;
-				resultExponentBits += (byte)(1 << MaximumExponentSize);
+				resultExponentBits += (1 << MaximumExponentSize);
 			}
 
-			return new Posit8_0(AssemblePositBitsWithRounding(resultSignBit, resultRegimeKValue, 0, resultFractionBits), true);
+			return new Posit8_0(AssemblePositBitsWithRounding(resultSignBit, resultRegimeKValue,(byte) 0, resultFractionBits), true);
 		}
 
 		public static Posit8_0 operator +(Posit8_0 left, int right) => left + new Posit8_0(right);
