+++ Posit/Posit8_2.cs
@@ -855,14 +855,14 @@ namespace Lombiq.Arithmetics
 			if (resultFractionBits == 0) return new Posit8_2(0, true);
 
 			var resultRegimeKValue = scaleFactor / (1 << MaximumExponentSize);
-			var resultExponentBits = (byte)(scaleFactor % (1 << MaximumExponentSize));
+			var resultExponentBits = (scaleFactor % (1 << MaximumExponentSize));
 			if (resultExponentBits < 0)
 			{
 				resultRegimeKValue -= 1;
-				resultExponentBits += (byte)(1 << MaximumExponentSize);
+				resultExponentBits += (1 << MaximumExponentSize);
 			}
 
-			return new Posit8_2(AssemblePositBitsWithRounding(resultSignBit, resultRegimeKValue, resultExponentBits, resultFractionBits), true);
+			return new Posit8_2(AssemblePositBitsWithRounding(resultSignBit, resultRegimeKValue,(byte) resultExponentBits, resultFractionBits), true);
 		}
 
 		public static Posit8_2 operator +(Posit8_2 left, int right) => left + new Posit8_2(right);
